update analytics chart helpers
